requirements.txt
boto3==1.17.89
botocore==1.20.89
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.17.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
